tests(compatibility): Babel-register shouldn't transpile the actual module.
Babel's require-hook was used for tests, because the tests were written
in ES2015+, but then it would also do a transpiling require on the
module. This meant the build transpile wasn't being adequately tested,
as a transpile would always occur during testing.

Fixed by limiting the scope of the require-hook.